requirements.txt
asgiref==3.5.2
django==4.1
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-sms==0.6.0
pillow==9.4.0
psycopg2==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2

